// src/app/displays/listado_Documento_Pendiente_Convertir/interfaces/print-convert.interface.ts
export interface PrintConvertInterface {
    monto_Letras?:                 string;
    id_Documento?:                 string;
    ref_Id_Documento?:             any;
    username?:                     string;
    razon_Social?:                 string;
    empresa_Nombre?:               string;
    empresa_Direccion?:            string;
    empresa_Nit?:                  string;
    pagina_Web?:                   string;
    empresa_Correo?:               string;
    empresa_PBX?:                  string;
    empresa_Direccion_F?:          string;
    empresa_Telefono?:             string;
    tipo_Documento?:               string;
    serie_Documento?:              string;
    serie_Resolucion?:             string;
    serie_Fecha_Autorizacion?:     string;
    serie_Ini?:                    string;
    serie_Fin?:                    string;
    serie_Forma_Pago?:             string;
    sucursal_Telefono?:            string;
    sucursal_Direccion?:           string;
    sucursal_Nombre?:              string;
    departamento_Serie?:           string;
    municipio_Serie?:              string;
    fE_Id_Documento?:              any;
    fE_Serie?:                     any;
    fE_CAE?:                       any;
    fE_FEcha?:                     any;
    documento_Nombre?:             string;
    documento_Direccion?:          string;
    documento_Nit?:                string;
    observacion_Documento?:        any;
    observacion_2?:                any;
    observacion_General?:          any;
    condiciones_Pago?:             number;
    documento_Telefono?:           string;
    documento_EMail?:              string;
    id_Cuenta?:                    string;
    fecha_Documento?:              Date;
    fecha_Hora?:                   Date;
    producto?:                     number;
    producto_Id?:                  string;
    unidad_Medida?:                string;
    simbolo?:                      string;
    des_Producto?:                 string;
    bodega?:                       string;
    cantidad?:                     number;
    cantidad_Pos?:                 number;
    observacion_1?:                string;
    monto_U_M_Tipo_Moneda?:        string;
    monto_Total_Tipo_Moneda?:      string;
    monto?:                        number;
    monto_Moneda?:                 number;
    moneda?:                       number;
    simbolo_Moneda?:               string;
    monto_Total_Unitario?:         number;
    monto_Total?:                  number;
    tipo_Cambio?:                  string;
    tipo_Documento_Personalizado?: string;
    direccion_Serie?:              string;
    telefono_Serie?:               string;
    atendio?:                      string;
    transaporte?:                  any;
    mensaje_Exento?:               string;
    oc?:                           any;
    op?:                           any;
    tipo_Factura?:                 any;
    leyenda?:                      any;
    regimen?:                      string;
    reimpresion?:                  string;
    sub_Total?:                    number;
    descuento?:                    number;
}